refactor: Merge duplicate batch-find-elements test files
- Merged batch-find-elements-integration.spec.ts and batch-find-elements-complex.spec.ts
- Eliminated ~19% code duplication
- Consolidated HTML templates and helper functions
- Organized tests into logical sections
- Reduced from 1627 lines across 2 files to 1570 lines in 1 file
