+++ src/diagnostics/element-discovery.ts
@@ -729,7 +729,7 @@ export class ElementDiscovery extends DiagnosticBase {
     return await element.evaluate((el) => {
       // Define regex inside evaluate context
       const CLASS_SPLIT_REGEX = /\s+/;
-      
+
       if (!(el instanceof Element)) {
         return 'unknown';
       }
